+++ R/cnt_skeleton.R
@@ -1,12 +1,23 @@
 #' Create a skeleton of a closed polygon object
 #'
 #' @param input \code{sf}, \code{sfc} or \code{SpatVector} polygons object
-#' @param simplify logical, indicating if the returning skeleton should be
-#' simplified using [rmapshaper::ms_simplify()] approach
-#' @param keep see below
-#' @param method see below
+#' @param keep numeric, proportion of points to retain (0.05-Inf; default 1).
+#' See Details.
 #'
-#' @inheritDotParams rmapshaper::ms_simplify
+#' @details
+#' - If \code{keep} equals 1 (default), no transformation will occur. The
+#' function will use the original geometry to find the skeleton.
+#'
+#' - If the keep parameter is below 1, then the [rmapshaper::ms_simplify()]
+#' function will be used with Douglas-Peuker algorithm. So the original input
+#' geometry would be simplified, and the resulting skeleton will be cleaner but
+#' maybe more edgy.
+#'
+#' - If the \code{keep} is above 1, then the densification
+#' algorithm is applied using the [geos::geos_densify()] function. This may
+#'  produce a very large object if keep is set more than 2. However, the
+#'  resulting skeleton would potentially be more accurate.
+#'
 #'
 #' @return An \code{sf}, \code{sfc} or \code{SpatVector} class
 #' object of a \code{LINESTRING} geometry
@@ -27,174 +38,287 @@
 #' plot(pol_skeleton)
 cnt_skeleton <-
   function(input,
-           simplify = TRUE,
-           keep = 0.1,
-           method = "dp",
-           ...) {
+           keep = 0.1) {
     UseMethod("cnt_skeleton")
   }
 
 #' @export
 cnt_skeleton.sf <-
   function(input,
-           simplify = TRUE,
-           keep = 0.1,
-           method = "dp",
-           ...) {
+           keep = 0.1) {
     cnt_skeleton_sf(
       input = input,
-      simplify = simplify,
-      keep = keep,
-      method = method,
-      ...
+      keep = keep
     )
   }
 
 #' @export
 cnt_skeleton.sfc <-
   function(input,
-           simplify = TRUE,
-           keep = 0.1,
-           method = "dp",
-           ...) {
-    cnt_skeleton_sf(
+           keep = 0.1) {
+    cnt_skeleton_sfc(
       input = input,
-      simplify = simplify,
-      keep = keep,
-      method = method,
-      ...
+      keep = keep
     )
   }
 
 #' @export
 cnt_skeleton.SpatVector <-
   function(input,
-           simplify = TRUE,
-           keep = 0.1,
-           method = "dp",
-           ...) {
+           keep = 0.1) {
     cnt_skeleton_terra(
       input = input,
-      simplify = simplify,
-      keep = keep,
-      method = method,
-      ...
+      keep = keep
     )
   }
 
 #' @export
 cnt_skeleton.character <-
   function(input,
-           simplify = TRUE,
-           keep = 0.1,
-           method = "dp",
-           ...) {
+           keep = 0.1) {
     cnt_skeleton_terra(
       input = input,
-      simplify = simplify,
-      keep = keep,
-      method = method,
-      ...
+      keep = keep
     )
   }
 
-cnt_skeleton_sf <-
+
+cnt_skeleton_sfc <-
   function(input,
-           simplify = TRUE,
-           keep = 0.1,
-           method = "dp",
-           ...) {
+           keep = 0.1) {
     # Check if input is of class 'sf' or 'sfc' and 'POLYGON'
     stopifnot(check_sf_polygon(input))
 
+    # Save CRS
     crs <-
       sf::st_crs(input)
 
-    pol <- input
+    # Transform to GEOS geometry
+    input_geos <-
+      input |>
+      geos::as_geos_geometry()
 
-    if (simplify) {
+    # Simplify or densify or do nothing
+    if (keep == 1) {
 
-      fixed_keep <-
-        ifelse(keep < 0.05, 0.05, keep)
+      pol_geos <- input_geos
 
-      pol <-
+    } else if (keep < 1 & keep >= 0.05) {
+
+      pol_geos <-
         rmapshaper::ms_simplify(
-          pol,
-          keep = fixed_keep,
-          method = method,
-          keep_shapes = T,
-          ...
+          input,
+          keep = keep,
+          method = "dp",
+          keep_shapes = TRUE
+        ) |>
+        geos::as_geos_geometry()
+
+    } else if (keep < 0.05) {
+
+      pol_geos <-
+        rmapshaper::ms_simplify(
+          input,
+          keep = 0.05,
+          method = "dp",
+          keep_shapes = TRUE
+        ) |>
+        geos::as_geos_geometry()
+
+    } else {
+
+      perimeter_length <-
+        geos::geos_length(input_geos)
+
+      point_count <-
+        input_geos |>
+        geos::geos_unique_points() |>
+        geos::geos_num_coordinates()
+
+      point_density <-
+        perimeter_length / point_count
+
+      pol_geos <-
+        geos::geos_densify(
+          input_geos,
+          tolerance = point_density / keep
         )
+
     }
 
-    pol_geos <-
-      geos::as_geos_geometry(pol)
-
+    # Find polygon skeleton
     pol_skeleton <-
       pol_geos |>
       geos::geos_voronoi_edges() |>
       geos::geos_intersection(pol_geos) |>
       geos::geos_unnest(keep_multi = F)
 
-    pol_sf <-
-      pol_skeleton |>
-      sf::st_as_sf()
-
+    # Transform back to sf
     pol_skeleton_crs <-
-      pol_sf |>
-      sf::st_set_crs(crs)
+      pol_skeleton |>
+      sf::st_as_sfc(crs = crs)
 
     return(pol_skeleton_crs)
   }
 
-
-cnt_skeleton_terra <-
+cnt_skeleton_sf <-
   function(input,
-           simplify = TRUE,
-           keep = 0.1,
-           method = "dp",
-           ...) {
-    # Check if input is of class 'SpatVector' and 'polygons'
-    stopifnot(check_terra_polygon(input))
+           keep = 0.1) {
+    # Check if input is of class 'sf' or 'sfc' and 'POLYGON'
+    stopifnot(check_sf_polygon(input))
 
-    pol <- sf::st_as_sf(input)
-
+    # Save CRS
     crs <-
-      sf::st_crs(pol)
+      sf::st_crs(input)
 
+    # Transform to GEOS geometry
+    input_geos <-
+      input |>
+      geos::as_geos_geometry()
 
-    if (simplify) {
+    # Simplify or densify or do nothing
+    if (keep == 1) {
 
-      fixed_keep <-
-        ifelse(keep < 0.05, 0.05, keep)
+      pol_geos <- input_geos
 
-      pol <-
+    } else if (keep < 1 & keep >= 0.05) {
+
+      pol_geos <-
         rmapshaper::ms_simplify(
-          pol,
-          keep = fixed_keep,
-          method = method,
-          keep_shapes = T,
-          ...
+          input,
+          keep = keep,
+          method = "dp",
+          keep_shapes = TRUE
+        ) |>
+        geos::as_geos_geometry()
+
+    } else if (keep < 0.05) {
+
+      pol_geos <-
+        rmapshaper::ms_simplify(
+          input,
+          keep = 0.05,
+          method = "dp",
+          keep_shapes = TRUE
+        ) |>
+        geos::as_geos_geometry()
+
+    } else {
+
+      perimeter_length <-
+        geos::geos_length(input_geos)
+
+      point_count <-
+        input_geos |>
+        geos::geos_unique_points() |>
+        geos::geos_num_coordinates()
+
+      point_density <-
+        perimeter_length / point_count
+
+      pol_geos <-
+        geos::geos_densify(
+          input_geos,
+          tolerance = point_density / keep
         )
+
     }
 
-    pol_geos <-
-      geos::as_geos_geometry(pol)
-
+    # Find polygon skeleton
     pol_skeleton <-
       pol_geos |>
       geos::geos_voronoi_edges() |>
       geos::geos_intersection(pol_geos) |>
       geos::geos_unnest(keep_multi = F)
 
-    pol_sf <-
-      pol_skeleton |>
-      sf::st_as_sf()
-
+    # Transform back to sf
     pol_skeleton_crs <-
-      pol_sf |>
-      sf::st_set_crs(crs) |>
-      terra::vect()
+      pol_skeleton |>
+      sf::st_as_sf(crs = crs) ##|>
+      # cbind(sf::st_drop_geometry(input))
+
+    return(pol_skeleton_crs)
+  }
+
+cnt_skeleton_terra <-
+  function(input,
+           keep = 0.1) {
+
+    # Check if input is of class 'SpatVector' and 'polygons'
+    stopifnot(check_terra_polygon(input))
+
+    # Save CRS
+    crs <-
+      terra::crs(input)
+
+    # Transform to GEOS geometry
+    input_geos <-
+      input |>
+      terra::geom(wkt = T) |>
+      geos::as_geos_geometry()
+
+    # Simplify or densify or do nothing
+    if (keep == 1) {
+
+      pol_geos <- input_geos
+
+    } else if (keep < 1 & keep >= 0.05) {
+
+      pol_geos <-
+        rmapshaper::ms_simplify(
+          sf::st_as_sf(input_geos),
+          keep = keep,
+          method = "dp",
+          keep_shapes = TRUE
+        ) |>
+        geos::as_geos_geometry()
+
+    } else if (keep < 0.05) {
+
+      pol_geos <-
+        rmapshaper::ms_simplify(
+          sf::st_as_sf(input_geos),
+          keep = 0.05,
+          method = "dp",
+          keep_shapes = TRUE
+        ) |>
+        geos::as_geos_geometry()
+
+    } else {
+
+      perimeter_length <-
+        geos::geos_length(input_geos)
+
+      point_count <-
+        input_geos |>
+        geos::geos_unique_points() |>
+        geos::geos_num_coordinates()
+
+      point_density <-
+        perimeter_length / point_count
+
+      pol_geos <-
+        geos::geos_densify(
+          input_geos,
+          tolerance = point_density / keep
+        )
+
+    }
+
+    # Find skeleton
+    pol_skeleton <-
+      pol_geos |>
+      geos::geos_voronoi_edges() |>
+      geos::geos_intersection(pol_geos) |>
+      geos::geos_unnest(keep_multi = F)
+
+    # Transform back to SpatVect
+    pol_skeleton_crs <-
+      pol_skeleton |>
+      wk::as_wkt() |>
+      as.character() |>
+      terra::vect(crs = crs)
+      # cbind(terra::as.data.frame(input))
 
     return(pol_skeleton_crs)
   }